Add action creator + called new reducer
